Update index.rst
In 'Configure the question block' changed "click" to "double-click" in two places for accuracy.
In 'Create a Script block' added a note to avoid copying/pasting the macro. Moved syntax statement to immediately below it for ease of reference.  In syntax description added "a single" in front of "tab character" to emphasize that only one should be used.